@@ -1,11 +1,11 @@
 name: listenable_list
 description: A new Flutter package project.
-version: 0.0.1
+version: 0.1.0
 homepage: https://github.com/JohnGu9
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
-  flutter: ">=1.17.0 <2.0.0"
+  sdk: '>=2.12.0-259.9.beta <3.0.0'
+  flutter: ">=1.24.0-10.2.pre <2.0.0"
 
 dependencies:
   flutter:
@@ -14,6 +14,7 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
+  test: ^1.16.4
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
